Move defaults to their own file
